Download Attachments From Multiple Emails VBA Code - Excel File
This is the Excel file to hold the VBA code (Download Attachments From Multiple Emails VBA Code - 12.01.2021.txt).

There are options for the drops down menus on the far right of the Sheet with the button.

To insert the code (if not there), go to the Developer tab and select Visual Basic from the ribbon. Then, add the code.